statd: not unlinking host files
Statd is not unlinking host files during SM_UNMON and
SM_UNMON_ALL calls because the given host is still on the run-time
notify list (rtnl) and the check flag is set when xunlink() is
called. But the next thing the caller of xunlink() does is
remove the host from the rtnl list which means the
unlink will never happen.

So this patch removes the check flag from xunlink() since
its not needed and correctly allocates and frees memory
used by xunlink().

Signed-off-by: Steve Dickson <steved@redhat.com>