configure: Add new build option "--enable-tirpc"
Allow easier testing of nfs-utils in legacy environments by providing
a "configure" option to force the build not to use libtirpc, even if
it's present on the build system.  This can also be tried as a
fallback if problems are found with the new TI-RPC-based nfs-utils
code.

The new option is:

  --enable-tirpc          enable use of TI-RPC [default=no]

Build option matrix:

   --disable-tirpc   --disable-ipv6	Default; IPv4 only, glibc RPC
   --disable-tirpc   --enable-ipv6	'configure' will fail
   --enable-tirpc    --disable-ipv6	IPv4 only, TI-RPC
   --enable-tirpc    --enable-ipv6	IPv4 and IPv6, TI-RPC

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>