Eliminate the displaying zero stats when the explicit protocol is specified (-2, -3, -4) the -Z and or --list options.
When a particular protocol is specified and either
the -Z or --list options are used, zeros or blank lines
are echoed to the screen when there is not any NFS traffic.
This cause any useful data to be scroll off the screen.

With this patch only non-zero stats will be shown, which
makes the output of these options more condensed and
in turn more useful.

Signed-off-by: Steve Dickson <steved@redhat.com>