mountd: move export lookup into separate function
Move this main loop to a separate function, to make it a little easier
to follow the logic of the caller.

Also, instead of waiting till we find an export to do the dns
resolution, do it at the start; it will normally be needed anyway, and
this simplifies the control flow.

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>