mountd: minor optimization in v4root_set
Since we're adding new exports as we traverse the export list, it's
possible we may find ourselves revisiting an export we just added.  It's
harmless to reprocess those exports, as we're currently doing.  But it's
also pointless.

(Actually, the current code appears to always add new export entries at
the head of each list, so we shouldn't hit this case.  It still may be a
good idea to keep this check, though, as insulation against future
changes to that data structure.)

Signed-off-by: J. Bruce Fields <bfields@citi.umich.edu>